Updated tests to use steps per entity
