+++ gamification-specs/src/main/resources/api-spec.yaml
@@ -67,6 +67,12 @@ paths:
           schema:
             type: string
           required: true
+        - in: query
+          name: usable
+          schema:
+            type: boolean
+            default: true
+          description: If the badge has to be desactivated pass this value with false
       requestBody:
         content:
           'application/json':
