+++ gamification-specs/src/test/java/ch/heigvd/amt/gamification/api/spec/helpers/Environment.java
@@ -1,14 +1,31 @@
 package ch.heigvd.amt.gamification.api.spec.helpers;
 
+import ch.heigvd.amt.gamification.ApiException;
+import ch.heigvd.amt.gamification.ApiResponse;
 import ch.heigvd.amt.gamification.api.DefaultApi;
+import lombok.Getter;
+import lombok.Setter;
 
 import java.io.IOException;
+import java.util.List;
 import java.util.Properties;
 import java.util.TimeZone;
 
 public class Environment {
 
     private DefaultApi api = new DefaultApi();
+    @Getter @Setter
+    ApiResponse lastApiResponse;
+    @Getter @Setter
+    ApiResponse lastApiResponseApp;
+    @Getter @Setter
+    ApiException lastApiException;
+    @Getter @Setter
+    boolean lastApiCallThrewException;
+    @Getter @Setter
+    int lastStatusCode;
+    @Getter @Setter
+    String lastReceivedLocationHeader;
 
     public Environment() throws IOException {
         Properties properties = new Properties();
@@ -22,6 +39,20 @@ public class Environment {
         return api;
     }
 
+    public void processApiResponse(ApiResponse apiResponse) {
+        lastApiResponse = apiResponse;
+        lastApiCallThrewException = false;
+        lastApiException = null;
+        lastStatusCode = lastApiResponse.getStatusCode();
+        List<String> locationHeaderValues = (List<String>)lastApiResponse.getHeaders().get("Location");
+        lastReceivedLocationHeader = locationHeaderValues != null ? locationHeaderValues.get(0) : null;
+    }
 
+    public void processApiException(ApiException apiException) {
+        lastApiCallThrewException = true;
+        lastApiResponse = null;
+        lastApiException = apiException;
+        lastStatusCode = lastApiException.getCode();
+    }
 
 }
